+++ changelog.md
@@ -4,6 +4,7 @@
 
 ### Patch
 
+- Updated dev dependencies.
 - Added a new [`hard-rejection`](https://npm.im/hard-rejection) dev dependency to ensure unhandled rejections in tests exit the process with an error.
 - Reorganized the test files.
 - Simplified `test/index.js`.