+++ changelog.md
@@ -8,6 +8,7 @@
 - Added a new [`hard-rejection`](https://npm.im/hard-rejection) dev dependency to ensure unhandled rejections in tests exit the process with an error.
 - Reorganized the test files.
 - Simplified `test/index.js`.
+- Reduced the `execute` middleware per request validation work.
 
 ## 4.0.0